group all \phantom related code together
